Migrate gorm to xorm refactor stage #1

+++ models/responses/contest_list.go
@@ -4,5 +4,5 @@ import "Rabbit-OJ-Backend/models"
 
 type ContestList struct {
 	List  []models.Contest `json:"list"`
-	Count uint32           `json:"count"`
+	Count int64            `json:"count"`
 }
