+++ models/responses/question_list.go
@@ -4,5 +4,5 @@ import "Rabbit-OJ-Backend/models"
 
 type QuestionListResponse struct {
 	List  []models.Question `json:"list"`
-	Count uint32            `json:"count"`
+	Count int64             `json:"count"`
 }
