+++ models/responses/submission_list.go
@@ -4,5 +4,5 @@ import "Rabbit-OJ-Backend/models"
 
 type SubmissionListResponse struct {
 	List  []models.SubmissionLite `json:"list"`
-	Count uint32                  `json:"count"`
+	Count int64                   `json:"count"`
 }
